requirements.txt
git+https://github.com/envinorma/envinorma-data.git@master
tqdm==4.30.0
pandas==1.2.4
python-swiftclient==3.11.1
python-keystoneclient==4.2.0
ocrmypdf==12.5.0
leginorma==0.0.3
text_diff==0.0.5
prefect==0.14.19